requirements.txt
asgiref==3.4.0
backports.entry-points-selectable==1.1.0
boto3==1.21.21
botocore==1.24.21
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.47.0
django-cors-headers==3.7.0
django-environ==0.6.0
django-filter==2.4.0
django-heroku==0.0.0
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
env==0.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
jmespath==1.0.0
Markdown==3.3.4
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
pandas==1.3.4
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
whitenoise==5.3.0
